+++ CHANGELOG.md
@@ -9,9 +9,21 @@
     Pay::Customer.where(processor: :paddle).update_all(processor: :paddle_classic)
     ```
 
-  Webhooks for `paddle.*` should be renamed to `paddle_classic.*`
+    Rename webhooks for `paddle.*` to `paddle_classic.*`
 
 * [Breaking] Subscriptions with `status: :canceled` and `ends_at: future` are now considered canceled. Previously, these were considered active to accomodate canceling a Braintree subscription during trial (and allowing the user to continue using until the end of the trial).
+* [Breaking] `stripe_account` has been moved from the `data:json` column to a dedicated column
+
+    To migrate, create a migration to add the column.
+    ```ruby
+    add_column :pay_customers, :stripe_account, :string
+    ```
+
+    Then copy the data over to the column:
+
+    ```ruby
+    Pay::Customer.find_each{ |c| c.update(stripe_account: c.data["stripe_account"]) }
+    ```
 
 ### 6.8.1